+++ src/main/java/com/ycloud/client/api/WhatsappPhoneNumbersApi.java
@@ -288,6 +288,150 @@ public class WhatsappPhoneNumbersApi {
         return new ApiListRequest();
     }
     /**
+     * Build call for register
+     * @param wabaId WhatsApp Business Account ID. (required)
+     * @param phoneNumber Phone number ID. (required)
+     * @param _callback Callback for upload/download progress
+     * @return Call to execute
+     * @throws ApiException If fail to serialize the request body object
+     * @http.response.details
+     <table summary="Response Details" border="1">
+        <tr><td> Status Code </td><td> Description </td><td> Response Headers </td></tr>
+        <tr><td> 200 </td><td> Successfully registered the phone number. </td><td>  -  </td></tr>
+        <tr><td> 404 </td><td> The requested resource does not exist. </td><td>  -  </td></tr>
+     </table>
+     */
+    public okhttp3.Call registerCall(String wabaId, String phoneNumber, final ApiCallback _callback) throws ApiException {
+        String basePath = null;
+        // Operation Servers
+        String[] localBasePaths = new String[] {  };
+
+        // Determine Base Path to Use
+        if (localCustomBaseUrl != null){
+            basePath = localCustomBaseUrl;
+        } else if ( localBasePaths.length > 0 ) {
+            basePath = localBasePaths[localHostIndex];
+        } else {
+            basePath = null;
+        }
+
+        Object localVarPostBody = null;
+
+        // create path and map variables
+        String localVarPath = "/whatsapp/phoneNumbers/{wabaId}/{phoneNumber}/register"
+            .replaceAll("\\{" + "wabaId" + "\\}", localVarApiClient.escapeString(wabaId.toString()))
+            .replaceAll("\\{" + "phoneNumber" + "\\}", localVarApiClient.escapeString(phoneNumber.toString()));
+
+        List<Pair> localVarQueryParams = new ArrayList<Pair>();
+        List<Pair> localVarCollectionQueryParams = new ArrayList<Pair>();
+        Map<String, String> localVarHeaderParams = new HashMap<String, String>();
+        Map<String, String> localVarCookieParams = new HashMap<String, String>();
+        Map<String, Object> localVarFormParams = new HashMap<String, Object>();
+
+        final String[] localVarAccepts = {
+            "application/json"
+        };
+        final String localVarAccept = localVarApiClient.selectHeaderAccept(localVarAccepts);
+        if (localVarAccept != null) {
+            localVarHeaderParams.put("Accept", localVarAccept);
+        }
+
+        final String[] localVarContentTypes = {
+            
+        };
+        final String localVarContentType = localVarApiClient.selectHeaderContentType(localVarContentTypes);
+        if (localVarContentType != null) {
+            localVarHeaderParams.put("Content-Type", localVarContentType);
+        }
+
+        String[] localVarAuthNames = new String[] { "api_key" };
+        return localVarApiClient.buildCall(basePath, localVarPath, "POST", localVarQueryParams, localVarCollectionQueryParams, localVarPostBody, localVarHeaderParams, localVarCookieParams, localVarFormParams, localVarAuthNames, _callback);
+    }
+
+    @SuppressWarnings("rawtypes")
+    private okhttp3.Call registerValidateBeforeCall(String wabaId, String phoneNumber, final ApiCallback _callback) throws ApiException {
+        
+        // verify the required parameter 'wabaId' is set
+        if (wabaId == null) {
+            throw new ApiException("Missing the required parameter 'wabaId' when calling register(Async)");
+        }
+        
+        // verify the required parameter 'phoneNumber' is set
+        if (phoneNumber == null) {
+            throw new ApiException("Missing the required parameter 'phoneNumber' when calling register(Async)");
+        }
+        
+
+        okhttp3.Call localVarCall = registerCall(wabaId, phoneNumber, _callback);
+        return localVarCall;
+
+    }
+
+    /**
+     * Register a phone number
+     * <p>
+     * Registers a WhatsApp business phone number.
+     * @param wabaId WhatsApp Business Account ID. (required)
+     * @param phoneNumber Phone number ID. (required)
+     * @return WhatsappPhoneNumber
+     * @throws ApiException If fail to call the API, e.g. server error or cannot deserialize the response body
+     * @http.response.details
+     <table summary="Response Details" border="1">
+        <tr><td> Status Code </td><td> Description </td><td> Response Headers </td></tr>
+        <tr><td> 200 </td><td> Successfully registered the phone number. </td><td>  -  </td></tr>
+        <tr><td> 404 </td><td> The requested resource does not exist. </td><td>  -  </td></tr>
+     </table>
+     */
+    public WhatsappPhoneNumber register(String wabaId, String phoneNumber) throws ApiException {
+        ApiResponse<WhatsappPhoneNumber> localVarResp = registerWithHttpInfo(wabaId, phoneNumber);
+        return localVarResp.getData();
+    }
+
+    /**
+     * Register a phone number
+     * <p>
+     * Registers a WhatsApp business phone number.
+     * @param wabaId WhatsApp Business Account ID. (required)
+     * @param phoneNumber Phone number ID. (required)
+     * @return ApiResponse&lt;WhatsappPhoneNumber&gt;
+     * @throws ApiException If fail to call the API, e.g. server error or cannot deserialize the response body
+     * @http.response.details
+     <table summary="Response Details" border="1">
+        <tr><td> Status Code </td><td> Description </td><td> Response Headers </td></tr>
+        <tr><td> 200 </td><td> Successfully registered the phone number. </td><td>  -  </td></tr>
+        <tr><td> 404 </td><td> The requested resource does not exist. </td><td>  -  </td></tr>
+     </table>
+     */
+    public ApiResponse<WhatsappPhoneNumber> registerWithHttpInfo(String wabaId, String phoneNumber) throws ApiException {
+        okhttp3.Call localVarCall = registerValidateBeforeCall(wabaId, phoneNumber, null);
+        Type localVarReturnType = new TypeToken<WhatsappPhoneNumber>(){}.getType();
+        return localVarApiClient.execute(localVarCall, localVarReturnType);
+    }
+
+    /**
+     * Register a phone number (asynchronously)
+     * <p>
+     * Registers a WhatsApp business phone number.
+     * @param wabaId WhatsApp Business Account ID. (required)
+     * @param phoneNumber Phone number ID. (required)
+     * @param _callback The callback to be executed when the API call finishes
+     * @return The request call
+     * @throws ApiException If fail to process the API call, e.g. serializing the request body object
+     * @http.response.details
+     <table summary="Response Details" border="1">
+        <tr><td> Status Code </td><td> Description </td><td> Response Headers </td></tr>
+        <tr><td> 200 </td><td> Successfully registered the phone number. </td><td>  -  </td></tr>
+        <tr><td> 404 </td><td> The requested resource does not exist. </td><td>  -  </td></tr>
+     </table>
+     */
+    public okhttp3.Call registerAsync(String wabaId, String phoneNumber, final ApiCallback<WhatsappPhoneNumber> _callback) throws ApiException {
+
+        okhttp3.Call localVarCall = registerValidateBeforeCall(wabaId, phoneNumber, _callback);
+        Type localVarReturnType = new TypeToken<WhatsappPhoneNumber>(){}.getType();
+        localVarApiClient.executeAsync(localVarCall, localVarReturnType, _callback);
+        return localVarCall;
+    }
+    /**
      * Build call for retrieve
      * @param wabaId WhatsApp Business Account ID. (required)
      * @param phoneNumber Phone number in [E.164](https://en.wikipedia.org/wiki/E.164) format. (required)